update user api done, check enum
