+++ backend/index.js
@@ -6,6 +6,7 @@ const dotenv = require('dotenv');
 const helmet = require('helmet');
 const morgan = require('morgan');
 const authRoute = require('./routes/auth');
+const userRoute = require('./routes/users');
 
 dotenv.config();
 
@@ -22,5 +23,6 @@ app.use(helmet());
 app.use(morgan('common'));
 
 app.use('/api/auth', authRoute);
+app.use('/api/users', userRoute);
 
 app.listen(port, () => { console.log(`Server listening to port ${port}`) });
